+++ src/features/index.js
@@ -1,3 +1,4 @@
 export { default as AgreementsAccordion } from './agreements-accordion';
 export { default as CustomLabelsAccordion } from './custom-labels-accordion';
 export { default as UsageConsolidationAccordion } from './usage-consolidation-accordion';
+export { default as ExportPackageResourcesModal } from './export-package-resources-modal';